+++ packages/state/index.ts
@@ -1,1 +1,1 @@
-export * from "./src/state";
+export {computed, signal, useEffect} from "./src/signals";
